move test and move its a svg test and drop 2 test
